+++ pytcl/core/maturity.py
@@ -303,6 +303,8 @@ MODULE_MATURITY: Dict[str, MaturityLevel] = {
     "dynamic_estimation.kalman.qmc": MaturityLevel.EXPERIMENTAL,  # new this release
     "dynamic_estimation.kalman.blue": MaturityLevel.EXPERIMENTAL,  # new this release
     "dynamic_estimation.batch_estimation": MaturityLevel.EXPERIMENTAL,  # new this release
+    "dynamic_estimation.batch_smoothers": MaturityLevel.EXPERIMENTAL,  # new this release
+    "dynamic_estimation.kalman.sqrt_cubature": MaturityLevel.EXPERIMENTAL,  # new this release
     # mathematical_functions
     "mathematical_functions.polynomials.multivariate": MaturityLevel.EXPERIMENTAL,  # new this release
     # static_estimation